consolidate handlebars
