--- .github/actions/build-ab/templates/index.handlebars
@@ -1,16 +1,0 @@
-// config
-window.NREUM={{{json config}}}
-// current
-{{{current}}}
-// next
-{{{next}}}
-// experiments
-{{#each experiments}}
-    // {{{this.name}}}
-    {{{this.contents}}}
-{{/each}}
-// post-script checks
-{{#each postScripts}}
-    // {{{this.name}}}
-    {{{this.contents}}}
-{{/each}}
